Updated providers and moved them into their own components, also created the providerselector component.
